添加全局错误提示

+++ Core.Model/ConfigModel/JwtAuthConfigModel.cs
@@ -11,10 +11,10 @@ namespace Core.Model.ConfigModel
             try
             {
                 JWTSecretKey = Configuration["JwtAuth:SecurityKey"];
-                WebExp = double.Parse(Configuration["JwtAuth:WebExp"]);
-                AppExp = double.Parse(Configuration["JwtAuth:AppExp"]);
-                MiniProgramExp = double.Parse(Configuration["JwtAuth:MiniProgramExp"]);
-                OtherExp = double.Parse(Configuration["JwtAuth:OtherExp"]);
+                WebExp = int.Parse(Configuration["JwtAuth:WebExp"]);
+                AppExp = int.Parse(Configuration["JwtAuth:AppExp"]);
+                MiniProgramExp = int.Parse(Configuration["JwtAuth:MiniProgramExp"]);
+                OtherExp = int.Parse(Configuration["JwtAuth:OtherExp"]);
             }
             catch (Exception e)
             {
@@ -29,18 +29,18 @@ namespace Core.Model.ConfigModel
         /// <summary>
         /// 
         /// </summary>
-        public double WebExp = 12;
+        public int WebExp = 12;
         /// <summary>
         /// 
         /// </summary>
-        public double AppExp = 12;
+        public int AppExp = 12;
         /// <summary>
         /// 
         /// </summary>
-        public double MiniProgramExp = 12;
+        public int MiniProgramExp = 12;
         /// <summary>
         /// 
         /// </summary>
-        public double OtherExp = 12;
+        public int OtherExp = 12;
     }
 }
